python-preprocessor/requirements.txt
python-pptx==0.6.21
flask==2.0.1
